feat: moved everything to a class
Also moved the script to a separate folder
